pyproject.toml
[tool.poetry]
name = "drf-course-work"
version = "0.1.0"
description = ""
authors = ["Sergius Art <xcxserj@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
ipython = "^8.18.1"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
celery = "^5.3.6"
redis = "^5.0.1"
django-celery-beat = "^2.5.0"
drf-yasg = "^1.21.7"
django-filter = "^23.4"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.3.1"
requests = "^2.31.0"
coverage = "^7.3.2"
flake8 = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
